Changes to pool size - Changed datatype to uint64 - Added/Implemented SetMax funtionality to PoolManager - Implemented the grpc and cli components for PoolSetMax - Removed unused pool min size

+++ server/server_test.go
@@ -100,7 +100,7 @@ func TestSrvPool(t *testing.T) {
 
 		poolCount := 5
 		pools := map[string]*api.Pool{}
-		poolMax := int32(10)
+		poolMax := uint64(10)
 		for i := 0; i < poolCount; i++ {
 			resp, err := client.PoolAdd(context.TODO(), &api.PoolAddRequest{
 				NetworkID:   networkResp.Network.ID,
